pyproject.toml
[project]
name = "wordle"
version = "0.1.0"
description = "A simple wordle clone with fasthtml"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml>=0.4.5",
]

# [tool.uv.sources]
# python-fasthtml = { git = "https://github.com/AnswerDotAI/fasthtml.git" }

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"


